refactor: 클래스 이름 및 위치 변경
